requirements.txt
flask==2.3.3
flask-sqlalchemy==3.1.1
flask-migrate==4.0.5
werkzeug==2.3.7
python-dotenv==1.0.0
pandas==2.1.1
numpy==1.25.2
gunicorn==21.2.0
firebase-admin==6.2.0
pyjwt==2.8.0
requests==2.31.0